+++ src/events/client/interactionCreate.ts
@@ -9,9 +9,7 @@ import {
 import { handleProfileInteraction } from "../../commands/game_commands/profile";
 import Session from "../../structures/session";
 import LocalizationManager from "../../helpers/localization_manager";
-
-export const BOOKMARK_COMMAND_NAME = "Bookmark Song";
-export const PROFILE_COMMAND_NAME = "Profile";
+import { BOOKMARK_COMMAND_NAME, PROFILE_COMMAND_NAME } from "../../constants";
 
 /**
  * Handles the 'interactionCreate' event
