+++ src/helpers/bot_listing_manager.ts
@@ -2,8 +2,8 @@ import Axios from "axios";
 import { IPCLogger } from "../logger";
 import State from "../state";
 import dbContext from "../database_context";
-import { VOTE_BONUS_DURATION } from "../commands/game_commands/vote";
 import { EnvType } from "../enums/env_type";
+import { VOTE_BONUS_DURATION } from "../constants";
 
 const logger = new IPCLogger("bot_stats_poster");
 interface BotListing {
