+++ src/structures/music_round.ts
@@ -1,5 +1,5 @@
 import type Eris from "eris";
-import { EMBED_SUCCESS_BONUS_COLOR } from "../helpers/discord_utils";
+import { EMBED_SUCCESS_BONUS_COLOR } from "../constants";
 import type PlayerRoundResult from "../interfaces/player_round_result";
 import type QueriedSong from "../interfaces/queried_song";
 import type UniqueSongCounter from "../interfaces/unique_song_counter";
