+++ src/structures/scoreboard.ts
@@ -4,8 +4,7 @@ import { getMention } from "../helpers/discord_utils";
 import type GuildPreference from "./guild_preference";
 import type SuccessfulGuessResult from "../interfaces/success_guess_result";
 import LocalizationManager from "../helpers/localization_manager";
-
-export const SCOREBOARD_FIELD_CUTOFF = 6;
+import { SCOREBOARD_FIELD_CUTOFF } from "../constants";
 
 export default class Scoreboard {
     /** Mapping of Discord user ID to Player */
