+++ src/test/ci/game_round.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/dot-notation */
 import assert from "assert";
-import { ExpBonusModifierValues } from "../../commands/game_commands/exp";
+import { ExpBonusModifierValues } from "../../constants";
 import { ExpBonusModifier } from "../../enums/exp_bonus_modifier";
 import { Gender } from "../../enums/option_types/gender";
 import { GuessModeType } from "../../enums/option_types/guess_mode_type";
