+++ src/test/ci/player.test.ts
@@ -1,5 +1,5 @@
 import assert from "assert";
-import { ExpBonusModifierValues } from "../../commands/game_commands/exp";
+import { ExpBonusModifierValues } from "../../constants";
 import { ExpBonusModifier } from "../../enums/exp_bonus_modifier";
 import Player from "../../structures/player";
